Extract Routstr wallet provider panel
